test/requirements.txt
pyaudio
websocket-client
python-dotenv
deepgram-sdk
openai

